Add support for Oracle Developer Studio C and MinGW-w64.
Tested with Oracle Developer Studio C 12.6 on Linux and Solaris.

Tested with MinGW-w64 on Fedora, Ubuntu, and Cygwin, targeting:

64-bit UCRT:   make CC=x86_64-mingw32ucrt-gcc
64-bit MSVCRT: make SIR_MSVCRT_MINGW=1 CC=x86_64-mingw32-gcc
32-bit MSVCRT: make SIR_MSVCRT_MINGW=1 CC=i686-w64-mingw32-gcc

+++ tests/tests.c
@@ -1538,7 +1538,8 @@ unsigned sirtest_thread(void* arg) {
 #endif
     }
 
-    printf("\thi, i'm thread id %d, logging to: '%s'...\n", threadid, my_args->log_file);
+    printf("\thi, i'm thread id %lld, logging to: '%s'...\n",
+            (long long)threadid, my_args->log_file);
 
     for (size_t n = 0; n < 1000; n++) {
         /* choose a random level, and colors. */
